terraform logic refactor to share variables

+++ terraform/modules/n8n/user_data.sh
@@ -1,5 +1,11 @@
-yum update -y
-yum install -y docker
-systemctl start docker
-systemctl enable docker
-docker-compose up -d
+#!/bin/bash
+
+if [ ! -f /var/log/first_boot_done ]; then
+
+    yum update -y
+    yum install -y docker
+    systemctl start docker
+    systemctl enable docker
+
+    touch /var/log/first_boot_done
+fi
